python api/requirements.txt
fastapi
uvicorn
requests
python-dotenv
python-multipart
langchain_community
langchain-text-splitters
langchain-huggingface
pypdf
sentence-transformers
faiss-cpu